+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 - Optimize allocations in Minimal Spanning Tree.
 - Export JuMP model from assigment problems.
 - Export JuMP model from zero-sum games.
+- Export JuMP model from Maximum Flow problems.
+- Export JuMP model from Minimum Cost Flow problems.